backend/services/python-services/requirements.txt
fastapi==0.115.12
MNN==3.1.3
numpy==1.26.4
opencv_python==4.11.0.86
Pillow==11.2.1
pydantic==2.11.5
ultralytics==8.3.102
uvicorn==0.34.3
python-multipart==0.0.20
httpx==0.27.0
kafka-python==2.2.13
tzlocal==5.0.1
